pyproject.toml
[tool.poetry]
name = "travel-blog"
version = "0.1.0"
description = ""
authors = ["rombaks"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
black = "^22.10.0"
isort = "^5.10.1"
django-filter = "^22.1"
Markdown = "^3.4.1"
flake8 = "^6.0.0"
mypy = "^0.991"
drf-yasg = "^1.21.4"
factory-boy = "^3.2.1"
Faker = "^15.3.4"
django-autoslug = "^1.9.8"

[tool.black]
line-length = 100
target-version = ['py39']
exclude = '''
(
  \.eggs
  |\.git
  |\.hg
  |\.mypy_cache
  |\.nox
  |\.tox
  |\.venv
  |_build
  |buck-out
  |build
  |dist
)
'''

[tool.isort]
line_length = 100
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
known_django = "django"
profile = "django"
src_paths = "app"
lines_after_imports = 2

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
